refactor: move API calls to services.py
